Move simulation to primitives
