+++ bridges/snowbridge/primitives/outbound-queue/src/v2/converter/shape.rs
@@ -1,8 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-FileCopyrightText: 2023 Snowfork <hello@snowfork.com>
 //! Structural validation shared with [`super::convert::XcmConverter::convert`] for early rejection
-//! (e.g. Bridge Hub Ethereum XCM simulation
-//! [`ShouldExecute`](xcm_executor::traits::ShouldExecute)).
+//! (e.g. Bridge Hub Ethereum XCM simulation in [`crate::v2::simulation::EthereumExportSimulationBarrier`]).
 
 use frame_support::{ensure, traits::ProcessMessageError};
 use snowbridge_core::AgentIdOf;
